+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-27T06:06:37.886Z",
+    "2026-03-27T12:10:41.461Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {