+++ packages/orq-rc/RELEASES.md
@@ -7799,3 +7799,13 @@ Based on:
 - [typescript v4.7.0-rc.23] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.23 - packages/orq-rc
+
+## 2026-03-27 12:13:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.24 - packages/orq-rc